Tests and fixes

+++ PackagePriceCalculator/Models/Pipe.cs
@@ -1,5 +1,5 @@
 ﻿namespace PackagePriceCalculator.Models;
 
-public class Pipe(int width, int height, int weight) : Package(width, height, weight)
+public class Pipe(int width, int height, double weight) : Package(width, height, weight)
 {
 }
